MAXX-32435 Accessibility for mixed Qt5 / legacy HWND hierarchies
Some QWidgets in 3ds Max are hosting native legacy HWNDs showing various UI
elements. To allow those hosted HWNDs to show up in the accessibility tree
used by automation, we check within the platform plugin, if the child accessible
found by Qt is hosting a native HWND. If so, we skip Qt's wrapped object and
jump straight to the native accessible retrieved from that hosted HWND.
As a result of this skipping, the hosting QWidget won't appear in the accessible
hierarchy at all - it gets transparently replaced by the hosted HWND.
In order to work as expected, the hosted HWND must therefore infill the hosting
QWidget completely.

The 3ds Max team has patched the QWinHost class of the QtWinMigrate project
also to use this property accordingly.