MAXX-34316 Fix mixture of native and non-native widgets (#4)
MAXX-34316 Fix mixture of native and non-native widgets

Since we have a lot of legacy code that rely on a fixed HWND parent hierarchy, we sometimes are forced to use a mixed native / non-native QWidget parent chain. In these cases the QWidget movement has to be propagated to the child chain, to ensure the correct positioning of the underlying native QWindow, even if the direct parent QWindow was not moved.