MAXX-36022 - QtDocking: Resizing side by side frames (#10)
* Autodesk 3ds Max addition: Extended docking resize behavior

The new 3ds Max docking resize behavior will work like this:

A drag move separator will now just do a single sided resizing of one
layout item and keep the size of the layout item on the other side of
the separator. The space that it needs for growing or shrinking will be
taken from the center docking area.

A shift+drag move separator will do the common Qt two sided resizing
where on both sides of the separator one item will grow and the other
one shrink. When the dragging is done in direction of the center docking
area and all items in that direction has been already shrunk to their
minimum size, then dragging doesn't get stuck as used to be, instead it
will continue and move the shrunken items into the center docking area.

* Minor improvements that came up in the code review.
- Put the new local methods in an anonymous namespace.
- some index bounds validations