add checks for resource creation failure for device lost
GPU crash is async and can happen at any timing. So it's necessary
to check all device resource creation. If fail, the code needs to
continue execution until Present() method to get the device removed,
lost or hung notification.

DXGI_ERROR_DEVICE_HUNG is also a GPU crash condition. Usually happens
when OGS execues a bad draw command, draw a very dense mesh, or issue
an instanced draw command with a lot of instances.

FLIP mode swap chain is not removed from HWND until flush. It needs
a call to device Flush() in order to create the new swap chain to
replace the old one. Otherwise, we will get access denied error
due to the old swap chain is not removed in time.

We need to reset the compositer on device lost. Because we need to
dispose all old resources. m_rhi is set to nullptr. On recovery,
a new rhi will be passed in. The resources will be re-created using
the new rhi.

Add check for create() methods and return on failure. The code needs
to continue to run to endFrame() so that it has a chance to know
if the render is device lost or not.