add more checks for resource creation failure for device lost
Swapchain can be nullptr in very rare case. We had already flushed
the command queue so the deferred deletion should happen already. But
it's always necessary to guard accessing nullptr. The next flush
will try to create the swapchain again.

m_widgetQuadData is created only once. It's very not likely to be
invalid. But just in case.