fix bug caused by https://git.autodesk.com/autodesk-forks/qtbase/commit/141d626029776a6b7a6f2fbc29a051221c0742e9
https://git.autodesk.com/autodesk-forks/qtbase/commit/141d626029776a6b7a6f2fbc29a051221c0742e9
This new commit sends WindowAboutToChange and WindowChange events to all
texture widgets in the toplevel window. When QQuickWidget receives the
WindowAboutToChange event, it will invalidate its render control and delete
its output texture. A dangling pointer is left in tlw's widgetTextures
list. The list is filled at the beginning of the paint cycle and used for
flushing to all native widgets in the tlw. If there is a dangling pointer
in the list, the next flush has a chance to crash. Since device lost will
not stop flushing. Qt will continue to loop over all native widgets and
flush.