FUS-209030 fix crash in d3d caused by dangling pointer on device lost & resize
If the window is resized after a device lost, Qt will recreate the
textures in new size but failed due to device lost situation.
A typical usage is to set the new size and call create(). Internally
in d3d11 RHI, the RHI object will call destroy() first and then create
the d3d11 object. In case of device lost, create() can fail and
return early. Now we have the following problems:

1) QD3D11TextureRenderTarget::d struct is not reset to nullptr in
destroy(). The class keeps a copy of rtv & dsv in d for use in
command buffer execution. If it's not cleared properly in destroy(),
OMSetRenderTargets will use dangling pointers to rtv & dsv because
the d3d objects are already released in destroy(). d only holds
a copy of the pointers.

2) QD3D11RenderBuffer::create() can release the d3d object without
increasing generation. In case of device lost, destroy() release
the object but create() returns earily. So generation is unchanged.
QD3D11TextureRenderTarget checks for generation change so it is
unware of the changes in QD3D11RenderBuffer. It holds a copy of
dsv pointer.

3) QD3D11Texture::create() and QD3D11Sampler::create() can release
the d3d object without increasing generation. In case of device
lost, destroy() release the object but create() returns earily. So
generation is unchanged. QD3D11ShaderResourceBindings checks for
generation change to decide whether to update its srv so it is
unaware of the changes in QD3D11Texture and QD3D11Sampler. It
holds a copy of srv pointer.