FUS-228518 more checks for resource creation failure on device lost
add more checks for resource creation failure for device lost

Swapchain can be nullptr in very rare case. We had already flushed the command queue so the deferred deletion should happen already. But it's always necessary to guard accessing nullptr. The next flush will try to create the swapchain again.

m_widgetQuadData is created only once. It's very not likely to be invalid. But just in case.

Change-Id: I96c0687f6d5be15b1cf626b718add70c0b7d8ed8