FUS-228522 fix dangling pointers in QWidgetRepaintManager::flush (#134)
The original Qt commit sends WindowAboutToChange and WindowChange events to all texture widgets in the toplevel window. When QQuickWidget receives the WindowAboutToChange event, it will invalidate its render control and delete its output texture. A dangling pointer is left in tlw's widgetTextures list. The list is filled at the beginning of the paint cycle and used for flushing to all native widgets in the tlw. If there is a dangling pointer in the list, the next flush has a chance to crash. Since device lost will not stop flushing. Qt will continue to loop over all native widgets and flush.

Change-Id: If3e69df4300bc6b0449db6b86602d91d05a3cd37